rpc: Fix named arguments in documentation
Summary:
```
This fixes a bug found with #18531:

    Currently the named argument for generateblock is documented as
address/descriptor, but the server only accepts a named argument of
address. Fix it by changing the name to output for both the
documentation and the server code. Also, add tests to prove the server
understands the new name output.

    Unrelated to that, there have been a bunch of aliases in the server
code that are not present in the source code of the documentation. Fix
that by adding the alias to the source code of the documentation. Only
the first alias is displayed in the rendered documentation. Also, add
tests to prove the server actually understands all aliases.
```

Backport of core [[https://github.com/bitcoin/bitcoin/pull/18607 | PR18607]].

Depends on D8567.

Test Plan:
  ninja all check-all

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Subscribers: deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D8568