Add upgraded nodes as seeds
Summary:
There's 100+ upgraded nodes on the network, so we should have this.  It should have been added for the 0.21.0 release,
otherwise early adopters get removed from the seeds list as we do new releases.

Test Plan: Run steps in `contrib/seeds/README.md` and the results should be mostly additive.

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D5724