[autopatch] Do not create a new git branch when fetching upstream
Summary:
This appears to be leftover from revisions of D4279. It's not clear if it's necessary
to name the local branch, but running the existing tests and poking around a bit points to no.
In fact, it causes confusion with the defaults set, as it creates a branch called "origin/master"
which then shadows the real origin/master. This patch fixes that behavior by removing the local
branch naming all together.

Test Plan:
Make sure tests still pass:
```
cmake -GNinja -DBUILD_SOURCE_CONTROL_TOOLS=ON ..
ninja check-source-control-tools
```

This should end in a detached HEAD state, not a branch named "origin/master":
```
./autopatch --dry-run --revision D5780
```

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D5786