[BENCH] Add an option to output the result as a Junit report
Summary:
This adds a Junit output option to the benchmark. This is expected to
allow for displaying graphs on CI to track the duration over time, and
eventually trigger an event if a diff makes this duration to incease.

Test Plan:
  ./src/bench/bitcoin-bench -printer=junit -filter="AES128CBC.*"
Check the output is a valid Junit XML.

Example result: https://build.bitcoinabc.org/viewLog.html?buildId=38288&buildTypeId=BitcoinABC_BitcoinAbcStaging&tab=testsInfo

Reviewers: #bitcoin_abc, jasonbcox

Reviewed By: #bitcoin_abc, jasonbcox

Subscribers: jasonbcox

Differential Revision: https://reviews.bitcoinabc.org/D5805