Add a deprecation notice for the autotools build system
Summary:
This add a mandatory option to run the `configure` script. The configure
stage will fail if the `--enable-deprecated-build-system` is not set,
and an error message instructs the user of the deprecation.

Depends on D5823.

Test Plan:
  ./configure
Check it fails early with the error message
  ./configure --enable-deprecated-build-system
  make
Check the build succeeds as usual.

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Subscribers: deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D5824