Rename IsGood() to IsReliable()
Summary:
`IsGood()` is a misleading name.  This function does not tell whether a
node is/was good or not. Rather, it uses the results of a
reliablity/survival function to inform the caller whether or not a node
has had reliable up-time.  As such `IsReliable()` more properly
describes to the function.

Test Plan:
  ninja
  ninja check-bitcoin-seeder

Reviewers: O1 Bitcoin ABC, #bitcoin_abc, jasonbcox

Reviewed By: O1 Bitcoin ABC, #bitcoin_abc, jasonbcox

Differential Revision: https://reviews.bitcoinabc.org/D5830