[CI] Run the tests for ARM
Summary:
The `qemu-user-static` allow the execution of the various non native
binaries via qemu transparently. We can take advantage of the cmake
crosscompiling emulator feature to wrap the calls to our binaries with
the proper qemu launcher and run the tests for non native platforms.

The ZMQ test is falling because zmq needs a feature unsupported by qemu,
so zmq is disabled for this build.

Depends on D5873.

Test Plan: Run the build-linux-arm CI build.

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Subscribers: deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D5893