Fix the emulator with autotools
Summary:
Accessing an undefined array key inside an `if` statement doesn't fail
the `if` but raises an exception. This is causing the python script
relying on this to check for the EMULATOR entry in the config file to
fail. This diff adds a dummy EMULATOR entry in the autotools config.ini
template to avoid this issue.

Test Plan:
  make check

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Subscribers: deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D5883