test: Pass ArgsManager into getarg_tests
Summary:
```
Replaced the global argsManager gArgs with a locally defined one in
getarg_tests. This is to avoid confusion in arg settings between the
test's ArgsManager and the #18804
```

Backport of core [[https://github.com/bitcoin/bitcoin/pull/18926 | PR18926]].

This overrides the changes from D4462 which achieved the same goal.

Test Plan:
  ninja check

Reviewers: #bitcoin_abc, majcosta

Reviewed By: #bitcoin_abc, majcosta

Differential Revision: https://reviews.bitcoinabc.org/D8643