Remove Core release note file, update previous release notes, and fix generate() deprecation message
Summary:
The core release note file was added rather than adding the release
notes to our file in D5971.  This also adjusts the version numbers to match our
own both in the release notes and in the rpc code.

Test Plan:
  ninja
  ./bitcoind
  ./bitcoin-cli generate 1
Verify the correct version numbers in the RPC deprecated message.

Reviewers: O1 Bitcoin ABC, #bitcoin_abc, jasonbcox

Reviewed By: O1 Bitcoin ABC, #bitcoin_abc, jasonbcox

Subscribers: jasonbcox

Differential Revision: https://reviews.bitcoinabc.org/D6074