Bump PORT_MIN in test framework to not collide with testnet
Summary:
As we add more functional tests, it becomes increasingly likely to see flaky failures if you
happen to be running a testnet node (default port 18333) while running functional tests on
the same machine. This is due to the p2p port range of the test suite being 16000 - 21000
which overlaps.  Although doing so is not recommended, there's no reason for us to knowingly
produce this flaky situation when all of these port numbers are under our control.

This patch bumps PORT_MIN so that the new rpc port range is 20000 - 25000 and the p2p port
range is 25000 - 30000.

Depends on D6075 just to bring us up-to-date with Core before making this small diversion.

Test Plan: `ninja check-functional`

Reviewers: #bitcoin_abc, Fabien, nakihito

Reviewed By: #bitcoin_abc, Fabien, nakihito

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D6076