Remove 0.20.x nodes from makeseeds
Summary:
With the upgrade imminent, 0.20.x nodes will not be in consensus so we should
no longer track them as seeds.

Note: As of writing this patch, this removes almost half of our seeds, but there's not much
we can do if those nodes do not upgrade.

Test Plan: Follow instructions in seeds README

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D6085