Check for IBD completion based on existing log message instead of 100% progress
Summary:
This is a more accurate test since 100% progress is actually post-IBD.

This will shorten the IBD test time as well as stabilize the amount of time it takes.
Currently, it's possible for a node to sit at 99.99% completion for tens of minutes
despite IBD being complete.

Test Plan: Run IBD on CI

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D6100