[trivial] Remove misleading 'const'
Summary:
Returning by const value is only meaningful in a specific circumstance around
user defined types. In this case, the const is not enforcing any restrictions
on the call site, so is misleading.

This is a backport of Core [[https://github.com/bitcoin/bitcoin/pull/18807 | PR18807]] [7/9]
https://github.com/bitcoin/bitcoin/pull/18807/commits/750456d6f29c63d57af05bfbdd6035bb9c965de2

Test Plan: ninja all check-all

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D9034