[CI] Disable unsupported qemu feature by bypassing the configuration
Summary:
This is a hack to allow for running the arm and aarch64 functional
tests.
Unsupported features are logged by qemu to stderr and are mixed with the
node stderr, causing the functional tests to fail.

Good news: qemu provides a lot of filtering and redirecting options for
its logs.
Bad news: the log filtering doesn't work at all (see bug
https://bugs.launchpad.net/qemu/+bug/1748612) so we can't use it.

For now the workaroung is to bypass the configuration detection to
disable the feature causing the log to appear.

Test Plan: Run the CI build-linux-arm and build-linux-aarch64 build plans.

Reviewers: #bitcoin_abc, jasonbcox

Reviewed By: #bitcoin_abc, jasonbcox

Differential Revision: https://reviews.bitcoinabc.org/D6203