refactor: Create interfaces earlier during initialization
Summary:
Add AppInitInterfaces function so wallet chain and chain client interfaces are
created earlier during initialization. This is needed in the next commit to
allow the gui splash screen to be able to register for wallet events through a
dedicated WalletClient interface instead managing wallets indirectly through
the Node interface. This only works if the wallet client interface is created
before the splash screen needs to use it.

This is a backport of [[https://github.com/bitcoin/bitcoin/pull/19099 | core#19099]] [1/3]
https://github.com/bitcoin/bitcoin/pull/19099/commits/b266b3e0bf29d0f3d5deaeec62d57c5025b35525

Test Plan: `ninja all check-all`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D10141