[CMAKE] Remove the ENABLE_WERROR option
Summary:
This option is confusing as it only turns a subset of warnings into
errors (some being only available on Clang), making it difficult to use.
It actually provides little to no benefit over the default ability to
add custom flags to the build, either with the environment `CFLAGS` and
`CXXFLAGS` or the cmake `CMAKE_C_FLAGS` and `CMAKE_CXX_FLAGS`.
Further more, `-Werror` is now enabled on CI (D6315).

Test Plan:
  cmake -GNinja .. -DENABLE_WERROR=ON
Check cmake warns about the unused variable.

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D6326