Introduce constant for mempool-based relay separate from mapRelay caching
Summary:
This constant is set to 2 minutes, rather than 15. This is still many times
larger than the transaction broadcast interval (2s for outbound, 5s for
inbound), so it should be acceptable for peers to know what our contents of
the mempool was that long ago.

This is a partial backport of [[https://github.com/bitcoin/bitcoin/pull/19109 | core#19109]] [2/5]
https://github.com/bitcoin/bitcoin/pull/19109/commits/b24a17f03982c9cd8fd6ec665b16e022374c96f0

Depends on D9434

Test Plan: `ninja all check-all`

Reviewers: #bitcoin_abc, majcosta

Reviewed By: #bitcoin_abc, majcosta

Subscribers: majcosta

Differential Revision: https://reviews.bitcoinabc.org/D9435