Rename Seeder's CAddrInfo -> CSeederAddrInfo
Summary:
The `CAddrInfo` is a class that already exists in the node software.
This will cause compiler issues as the seeder is built with more parts
of the node software. Because the seeder does not record the same
information about an address as the node software, refactoring the
seeder's version of the class to inherit the node software's version
would not be useful.

scripted diff
`sed -i -e 's/CAddrInfo/CSeederAddrInfo/g' src/seeder/db.*`

Test Plan:
  ninja

Reviewers: O1 Bitcoin ABC, #bitcoin_abc, jasonbcox

Reviewed By: O1 Bitcoin ABC, #bitcoin_abc, jasonbcox

Subscribers: jasonbcox

Differential Revision: https://reviews.bitcoinabc.org/D6355