refactor: Get rid of more redundant chain methods
Summary:
This just drops three `interfaces::Chain` methods replacing them with other calls.

Motivation for removing these chain methods:

- Need to get rid of findFirstBlockWithTimeAndHeight for [[ https://github.com/bitcoin/bitcoin/pull/10102 | #10102 ]], which doesn't
  support overloaded methods
- Followup from
  https://github.com/bitcoin/bitcoin/pull/16426#discussion_r412487403
- phantomcircuit comments about findNextBlock test
  http://www.erisian.com.au/bitcoin-core-dev/log-2020-06-06.html#l-214

Behavior is not changing in any way here. A TODO comment in
ScanForWalletTransactions was removed, but just because it was invalid (see
https://github.com/bitcoin/bitcoin/pull/19195#discussion_r448020762), not
because it was implemented.

This is a backport of [[https://github.com/bitcoin/bitcoin/pull/19425 | core#19425]] [1/3]
https://github.com/bitcoin/bitcoin/pull/19425/commits/3fbbb9a6403a86fbed3d5d9f7939998922593377

Test Plan: `ninja all check-all`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D11418