[CMAKE] Don't require jemalloc for systems where it's the default
Summary:
On these system there is no need to pull jemalloc as a 3rd party lib
since it's the built-in default. This fixes the build on FreeBSD.

Test Plan:
Build on Linux and check jemalloc is required by default.
  cmake -GNinja .. -DUSE_JEMALLOC=OFF
Check disabling jemalloc manually still works as expected.

Build on FreeBSD and check the build succeeds (tested without Qt as per our documentation).

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D6544