[Qt] Fix deprecated QDateTime(const QDate &)
Summary:
Qt 5.15 has deprecated the construction of a `QDateTime` object from a
`QDate`. This was used in our codebase to build a `QDateTime`
corresponding to the start of the day for the given `QDate`, but Qt
introduced a dedicated method for this use case since 5.14.

Test Plan:
With Qt 5.5.1 and Qt 5.15:
  ninja all check

Run bitcoin-qt and check the transaction view filter still works as
expected (including with an arbitrary range input).

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D6575