[refactor] Add deploymentstatus.h
Summary:
Provides DeploymentEnabled, DeploymentActiveAt, and DeploymentActiveAfter helpers for checking the status of buried deployments.

This is a partial backport of [[https://github.com/bitcoin/bitcoin/pull/19438 | core#19438]]
https://github.com/bitcoin/bitcoin/pull/19438/commits/2b0d291da8f479739ff394dd92801da8c40b9f8e#

Note that `GetBlockScriptFlags` is off-by-one when applied to mempool transaction in the source material, hence the use of `DeploymentActiveAfter` instead of `DeploymentActiveAt` in our codebase (see explanation in D3361).

Depends on D12219

Test Plan:
`ninja all check-all`

Complete an IBD.

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D12220