wallet: Replace CDataStream& with CDataStream&& where appropriate
Summary:
The keys and values are only to be used once because their memory is set
to zero. Make that explicit by moving the bytes into the lower level
methods.

This is a backport of [[https://github.com/bitcoin/bitcoin/pull/19320 | core#19320]] [2/2]
https://github.com/bitcoin/bitcoin/pull/19320/commits/fa8a341b88cabfd7f8d702db7cb9972b0804bf2a

Depends on D9982

Test Plan: `ninja all check-all`

Reviewers: #bitcoin_abc, majcosta

Reviewed By: #bitcoin_abc, majcosta

Subscribers: majcosta

Differential Revision: https://reviews.bitcoinabc.org/D9983