[CI] Convert the build-configurations.sh script to python, read config
Summary:
This diff converts the `build-configurations.sh` script to python, and
make it read the build configuration from a json file. For now it is
limited to running the script and forwarding the error code.

This configuration file will later allow to add more features.
The bash script is kept as a dummy wrapper to keep compatibility with
existing infra.

Test Plan: Run any CI build, both with the python and the shell script.

Reviewers: #bitcoin_abc, jasonbcox

Reviewed By: #bitcoin_abc, jasonbcox

Subscribers: jasonbcox

Differential Revision: https://reviews.bitcoinabc.org/D6711