Revert "[CI] Install the latest wine version from the winehq repository"
Summary:
This reverts commit b25c0bf30ae0e20e5bc506bcc07e2b96934dedd5.

The `libfaudio0` dependency installation is subject to change for every
wine update, the maintenance cost is not worth it.

Another solution needs to be found for running the latest `wine`. One of
the options is to use a debian testing container that provides
`libfaudio0`. Package pinning is not an option here as it requires an
update from the libc.

Test Plan: Run the CI build-win64 build plan.

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D6676