[CI] Don't print an error if there is no sanitizer log files
Summary:
This prints a confusing error message when there is no sanitizer log
(e.g. in case the failure is not due to the sanitizer). This diff fixes
the issue by testing for the existance of the file prior to printing it.

Test Plan:
$
Edit the build configuration so that the print_sanitizers_log traps the
EXIT instead of ERR. Run any CI build plan and check that the following
error message is no longer displayed:
  cat: '/tmp/sanitizer_logs/*.log.*': No such file or directory

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D6680