[CI] Extract the CI build facitilties to it's own script and source it
Summary:
This is a step toward splitting the `build-configurations.sh` into
separated scripts that can be individually configured.
This is move only, no change in behavior.

Note: `shellcheck` needs to be told how to search for the sourced file
in order to parse it.

Test Plan: Run any CI build plan.

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D6701